% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSummary_.R
\name{AccountSummary_}
\alias{AccountSummary_}
\title{Return the investor's Account Summary}
\usage{
AccountSummary_(LC_CRED = NULL)
}
\arguments{
\item{LC_CRED}{Credential object having class LendingClub_credential. This object can be created using the \code{MakeCredential()} function. You can pass the object directly as an argument to the function or alternatively, it can be read from the global environment if it was stored as "LC_CRED".}
}
\value{
Object of class LendingClub_api. The object has two components. The first is \code{$content} with a data.frame of attributes of the invenstor's account such as available cash and value of the portfolios. The attributes are arranged by rows in the first column and the values are in the second column.  The second component is the \code{$response} including a timestamp of when the data was retrieved.
}
\description{
Access the user's account and return the summary data.  The data points
include available cash, total account balance, and the number of notes in the
portfolio.
}
\examples{
\dontrun{
LC_CRED<- MakeCredential(InvestorID, APIKey)
AccountSummary()}

}
