% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvailableCash.R
\name{AvailableCash}
\alias{AvailableCash}
\title{Return Investor's available cash}
\usage{
AvailableCash()
}
\arguments{
\item{investorID}{Investor's ID number from the Account Summary page}

\item{key}{Investor's API key. See the Account Settings page}
}
\description{
Uses Lending Club's API to access the user's account and return the available
 cash.
}

