% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuySecondary.R
\name{BuySecondary}
\alias{BuySecondary}
\title{Buy a note from the secondary market}
\usage{
BuySecondary(loanId, orderId, noteId, bidPrice)
}
\arguments{
\item{loanId}{Numeric identifier of the loan whose note requested for purchase}

\item{orderId}{Numeric identifier of the order for the note}

\item{noteId}{Numeric identifier of the note}

\item{bidPrice}{Positive numeric value in dollars ($) and cents representing the buy price desired for the note}
}
\description{
Buy a note from the secondary market
}

