% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{CleanData}
\alias{CleanData}
\title{Convert the list to dataframe}
\usage{
CleanData(x, ncols)
}
\arguments{
\item{x}{list to be cleaned}
}
\description{
Converts the list of notes from function NotesOwned to a dataframe
}

