% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanNoteData.R
\name{CleanNoteData}
\alias{CleanNoteData}
\title{Convert the list to dataframe}
\usage{
CleanNoteData(x)
}
\arguments{
\item{x}{list to be cleaned}
}
\description{
Converts the list of notes from function NotesOwned to a dataframe
}

