% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePortfolio.R
\name{CreatePortfolio}
\alias{CreatePortfolio}
\title{Create a new portfolio}
\usage{
CreatePortfolio(port_name, port_desc)
}
\arguments{
\item{port_name}{Name of the new portfolio}

\item{port_desc}{Portfolio description}
}
\description{
Uses Lending Club's API to create a new portfolio
}

