% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetailedNotesOwned.R
\name{DetailedNotesOwned}
\alias{DetailedNotesOwned}
\title{Return detail on investor's loans}
\usage{
DetailedNotesOwned()
}
\description{
Uses Lending Club's API to access the user's account and return
detail on the notes owned by the investor
}

