% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolioBuy.R
\name{FolioBuy}
\alias{FolioBuy}
\title{Buy notes from the secondary market}
\usage{
FolioBuy(loanId, orderId, noteId, bidPrice)
}
\arguments{
\item{loanId}{Loan ID number}

\item{orderId}{Order ID number}

\item{noteId}{Note Id number}

\item{bidPrice}{Notes list price.}
}
\description{
Buy notes from the folio secondary market.
}

