% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolioOrders.R
\name{FolioOrders}
\alias{FolioOrders}
\title{Folio orders submitted in the current month}
\usage{
FolioOrders(includeDetails = TRUE, orderType, status = NULL,
  LC_CRED = NULL)
}
\arguments{
\item{includeDetails}{Include note level details}

\item{orderType}{Return 'buy' or 'sell' orders}

\item{status}{Used to filter results. Can be NULL or one of 9 possible status:
\itemize{
\item IN_LISTING
\item SETTLED
\item CANCELED_BY_SELLER
\item CANCELED_TIME_EXPIRED
\item PENDING_TRADE_SETTLEMENT
\item CANCELED_BY_PAYMENT
\item CANCELED_BY_DEFAULT_OR_CHARGE_OFF
\item REPRICED
\item CANCELED_BY_BANKRUPTCY
}}

\item{LC_CRED}{Credential object having class LendingClub_credential. This object can be created using the \code{MakeCredential()} function. You can pass the object directly as an argument to the function or alternatively, it can be read from the global environment if it was stored as "LC_CRED".}
}
\description{
View either buy or sell orders placed in the current month
}
