% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolioSell.R
\name{FolioSell}
\alias{FolioSell}
\title{Sell notes}
\usage{
FolioSell(loanId, orderId, noteId, expireDate = NULL, askingPrice)
}
\arguments{
\item{loanId}{The loan ID}

\item{orderId}{the loan's order ID}

\item{noteId}{The loan's note ID}

\item{expireDate}{the date the sell offer will expire. Can be as short as the
current date or up to 7 days from posting the offer. Character string.}

\item{askingPrice}{Price you wish to sell your loan for. Remember to consider
the accrued interest in your offering.}
}
\description{
Sell notes on the FOLIOfn secondary market
}

