% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_Dashboard.R
\name{lc_Dashboard}
\alias{lc_Dashboard}
\title{Creates an HTML dashboard}
\usage{
lc_Dashboard(filename = "Dashboard.html", out_dir = getwd(),
  ID = LC_CRED$investorID, key = LC_CRED$key, ...)
}
\arguments{
\item{filename}{Name of the file to create}

\item{out_dir}{Directory where to save the html file}

\item{ID}{User's ID from the Account Summary page}

\item{key}{User's API key from the account settings page}

\item{...}{Arguments passed on to rmarkdown::render}
}
\description{
Knits a document with summary statistics needed to evaluate an account.
}
