% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC_GET.R
\name{LC_GET}
\alias{LC_GET}
\title{Wrapper for GET}
\usage{
LC_GET(searchURL, auth, query = NULL)
}
\arguments{
\item{searchURL}{URL passed to the Lending Club API}

\item{auth}{API key}
}
\description{
A wrapper for the GET function to pull the needed data
}

