% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC_POST.R
\name{LC_POST}
\alias{LC_POST}
\title{Wrapper for POST}
\usage{
LC_POST(postURL, params, auth)
}
\arguments{
\item{postURL}{URL passed to the Lending Club API}

\item{params}{parameters passed to the API}

\item{auth}{API key}
}
\description{
A wrapper for POST which passes the data to the L/C server. Data
is encoded as json.
}
\keyword{internal}
