% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClub-deprecated.R
\docType{package}
\name{LendingClub-deprecated}
\alias{LendingClub-deprecated}
\alias{AccountSummary_}
\alias{AddFunds}
\alias{AvailableCash}
\alias{CancelTransfer}
\alias{CreatePortfolio}
\alias{MakeCredential}
\alias{LendingClub-deprecated-package}
\title{Deprecated functions in the \pkg{LendingClub} package.}
\usage{
AccountSummary_(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern versions of the function}
}
\description{
The functions listed below are provided for compatability with older
versions of the LendingClub package. They are deprecated and may
be removed (defunct) in later versions.
\itemize{
\item \code{ \link{AccountSummary}} is now \code{ \link{lc_AccountSummary}}
\item \code{ AccountSummary_} is now \code{ \link{lc_AccountSummary}}
\item \code{ AddFunds} is now \code{\link{lc_AddFunds}}
\item \code{ AvailableCash} is now \code{\link{lc_AvailableCash}}
\item \code{ CreatePortolio} is now \code{\link{lc_CreatePortfolio}}
\item \code{ MakeCredential} is now \code{\link{lc_MakeCredential}}
\item \code{ Dashboard} is now \code{\link{lc_Dashboard}}
\item \code{ DetailedNotesOwned} is now \code{\link{lc_DetailedNotesOwned}}
\item \code{ NotesOwned } is now \code{\link{lc_NotesOwned}}
}
}
