% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClub.R
\docType{package}
\name{LendingClub}
\alias{LendingClub}
\alias{LendingClub-package}
\title{Lending Club}
\usage{
vignette("LendingClub")
}
\description{
To create the tools for working with the Lending Club API.
}
\author{
Ryan Kuhn, CPA
}

