\name{LendingClub-package}
\alias{LendingClub-package}
\alias{LendingClub}
\docType{package}
\title{
\packageTitle{LendingClub}
}
\description{
\packageDescription{LendingClub}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LendingClub}
\packageIndices{LendingClub}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LendingClub}

Maintainer: \packageMaintainer{LendingClub}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
