% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeURL.R
\name{MakeURL}
\alias{MakeURL}
\title{Make the search URL}
\usage{
MakeURL(investorID, append)
}
\arguments{
\item{investorID}{Investor ID which is the same as the account number}

\item{append}{string to append at the end of the URL}
}
\description{
Assembles the search URL from the invstor ID and the type of query
}
\keyword{internal}

