% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PendingTransfers.R
\name{PendingTransfers}
\alias{PendingTransfers}
\title{Return Investor's Pending Transfers}
\usage{
PendingTransfers()
}
\description{
Uses Lending Club's API to access the user's account and return the summary
}

