% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{print.AccountSummary}
\alias{print.AccountSummary}
\title{Print the output from an API call}
\usage{
\method{print}{AccountSummary}(x, ...)
}
\arguments{
\item{x}{an object of class LendingClub_api}

\item{...}{To pass additional arguments to the generic print function}
}
\description{
This is a general function to print the content from an API
call. The API call returns a list of 2 elements: The first
is the content from the call and the second element is the
actual response.
}
