% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{print.LendingClub_api}
\alias{print.LendingClub_api}
\title{Print the Output from an API call}
\usage{
print(object)
}
\arguments{
\item{object}{an object of class LendingClub_api}
}
\description{
This is a general function to print the conent from an API
call. The API call returns a list of 2 elements: The first
is the content from the call and the second element is the
actual response.
}

