% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{print.LendingClub_credential}
\alias{print.LendingClub_credential}
\title{Print the LC_CRED object}
\usage{
\method{print}{LendingClub_credential}(x, ...)
}
\arguments{
\item{x}{an object of class LendingClub_api}

\item{...}{To pass additional arguments to the generic print function}
}
\description{
This function prints the content from the MakeCredential 
call.
}
