% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondaryListings.R
\name{SecondaryListing}
\alias{SecondaryListing}
\title{Loans listed on Folio marketplace}
\usage{
SecondaryListing(updatedSince = 5)
}
\arguments{
\item{updatedSince}{filter listings to include only ones that were modified/created within the last X minutes.}
}
\description{
Returns a dataframe of loans listed on the secondary market.
}
