% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summmary.R
\name{summary.holdings}
\alias{summary.holdings}
\title{Prepare a dashboard of key metrics}
\usage{
\method{summary}{holdings}(object, ...)
}
\arguments{
\item{object}{data.frame of loans from DetailedNotesOwned()}

\item{...}{currently not used}
}
\description{
Creating a summary of matrics an investor would find useful in evaluating their
account.
}
\details{
At Risk Ratio calculates the percentage of the loans with principal 16 or more days late.
}
\examples{
\dontrun{
summary( DetailedNotesOwned( LC_CRED ))

# split notes on a variable and  run summary for each subset
holdings<- DetailedNotesOwned(LC_CRED)$content
x<- lapply(split(holdings, f= holdings$portfolioName), summary)
y<- Reduce(function(...) merge(..., by= 'col_labs',  all.x=TRUE), x)
names(y)[-1]<- names(x)
y
}
}
