% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFunds.R
\name{TransferFunds}
\alias{TransferFunds}
\title{Setup a funds transfer}
\usage{
TransferFunds(amount, freq = NULL, start = NULL, end = NULL)
}
\arguments{
\item{amount}{amount to be transferred}

\item{freq}{frequency of transfers. Optional.  If a value is not
supplied, then the transfer will be processed immediately.}

\item{start}{date formatted as YYYY/MM/DD. If a date is not}

\item{end}{date formatted as YYYY/MM/DD}
}
\description{
Sets up a recurring or one-time funds transfer. The investor must
already have setup the account from which the funds are transferred.
Transfers can one-time or recurring.
}

