% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithdrawFunds.R
\name{WithdrawFunds}
\alias{WithdrawFunds}
\title{Withdraw funds}
\usage{
WithdrawFunds(amount)
}
\arguments{
\item{amount}{amount to be transferred}
}
\description{
Withdraw funds from the investor's account
}

