% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSummary.R
\name{AccountSummary}
\alias{AccountSummary}
\title{Return Investor's Account Summary}
\usage{
AccountSummary()
}
\description{
Uses Lending Club's API to access the user's account and return the summary
}

