% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSummary.R
\name{AccountSummary}
\alias{AccountSummary}
\title{Return the Investor's Account Summary}
\usage{
AccountSummary(LC_CRED = NULL)
}
\arguments{
\item{LC_CRED}{Credential object having class LendingClub_credential. This object can be created using the \code{MakeCredential()} function. You can pass the object directly as an argument to the function or alternatively, it can be read from the global environment if it was stored as "LC_CRED".}
}
\description{
Access the user's account and return the summary data.  The data points
include available cash, total account balance, and the number of notes in the
portfolio.
}

