% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvailableCash.R
\name{AvailableCash}
\alias{AvailableCash}
\title{Return Investor's available cash}
\usage{
AvailableCash()
}
\description{
Uses Lending Club's API to access the user's account and return the available
cash.
}

