% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetailedNotesOwned.R
\name{DetailedNotesOwned}
\alias{DetailedNotesOwned}
\title{Return detail on investor's loans}
\usage{
DetailedNotesOwned(LC_CRED = NULL)
}
\arguments{
\item{LC_CRED}{Credential object having class LendingClub_credential. This object can be created using the \code{MakeCredential()} function. You can pass the object directly as an argument to the function or alternatively, it can be read from the global environment if it was stored as "LC_CRED".}
}
\description{
Get a detailed listing of notes owned by the investor.
}

