% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolioBuy.R
\name{FolioBuy}
\alias{FolioBuy}
\title{Buy notes from the secondary market}
\usage{
FolioBuy(loanId, orderId, noteId, bidPrice, LC_CRED = NULL)
}
\arguments{
\item{loanId}{Loan ID number}

\item{orderId}{Order ID number}

\item{noteId}{Note Id number}

\item{bidPrice}{Notes list price.}

\item{LC_CRED}{Credential object having class LendingClub_credential. This object can be created using the \code{MakeCredential()} function. You can pass the object directly as an argument to the function or alternatively, it can be read from the global environment if it was stored as "LC_CRED".}
}
\description{
Buy notes from the folio secondary market.
}

