% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolioBuy.R, R/FolioSell.R
\name{FolioBuy}
\alias{FolioBuy}
\alias{FolioSell}
\title{Transact with the secondary market}
\usage{
FolioBuy(loanId = NULL, orderId = NULL, noteId, Price = NULL,
  LC_CRED = NULL, quiet = T)

FolioSell(loanId = NULL, orderId = NULL, noteId, expireDate = NULL,
  Price = NULL, LC_CRED = NULL, quiet = T)
}
\arguments{
\item{loanId}{\strong{\emph{Deprecated}}.  As of Jan 2018, the folio API was revised and the loan ID is no longer required.}

\item{orderId}{\strong{\emph{Deprecated}}.  As of Jan 2018, the folio API was revised and the order ID is no longer required.#' @inheritParams AccountSummary}

\item{noteId}{The loan's note ID}

\item{Price}{Price you wish to sell your loan for. Remember to consider}

\item{quiet}{Should the API response be printed to the console
the accrued interest in your offering.}

\item{expireDate}{The date the sell offer will expire. Can be as short as the
current date or up to 7 days from posting the offer. Character string.}
}
\description{
Buy or sell notes on the FOLIOfn secondary market
}
