% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolioBuy.R, R/FolioSell.R
\name{FolioBuy}
\alias{FolioBuy}
\alias{FolioSell}
\title{Transact with the secondary market}
\usage{
FolioBuy(loanId, orderId, noteId, Price = NULL, LC_CRED = NULL, quiet = T)

FolioSell(loanId, orderId, noteId, expireDate = NULL, Price = NULL,
  LC_CRED = NULL, quiet = T)
}
\arguments{
\item{loanId}{The loan ID}

\item{orderId}{The loan's order ID}

\item{noteId}{The loan's note ID}

\item{Price}{Price you wish to sell your loan for. Remember to consider}

\item{LC_CRED}{Credential object having class LendingClub_credential. This object can be created using the \code{MakeCredential()} function. You can pass the object directly as an argument to the function or alternatively, it can be read from the global environment if it was stored as "LC_CRED".}

\item{quiet}{Should the response be printed to the console
the accrued interest in your offering.}

\item{expireDate}{The date the sell offer will expire. Can be as short as the
current date or up to 7 days from posting the offer. Character string.}
}
\description{
Buy or sell notes on the FOLIOfn secondary market
}
