% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClub-deprecated.R
\name{LendingClub-deprecated}
\alias{LendingClub-deprecated}
\title{Deprecated functions in package \pkg{LendingClub}.}
\description{
The functions listed below are deprecated and will be removed (defunct) in
the near future.
\itemize{
\item \code{ \link{AccountSummary}}, use \code{ \link{AccountSummary_}}
}
}
