% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClub-deprecated.R
\docType{package}
\name{LendingClub-deprecated}
\alias{LendingClub-deprecated}
\alias{AccountSummary_}
\alias{AddFunds}
\alias{AvailableCash}
\alias{CancelTransfer}
\alias{CreatePortfolio}
\alias{MakeCredential}
\alias{Dashboard}
\alias{DetailedNotesOwned}
\alias{NotesOwned}
\alias{PendingTransfers}
\alias{SubmitOrder}
\alias{ListedLoans}
\alias{PortfoliosOwned}
\alias{WithdrawFunds}
\alias{LendingClub-deprecated-package}
\title{Deprecated functions in the \pkg{LendingClub} package.}
\arguments{
\item{...}{Parameters to be passed to the modern versions of the function}
}
\description{
The functions listed below are provided for compatability with older
versions of the LendingClub package. They are deprecated and may
be removed (defunct) in later versions.
}
\details{
\tabular{ll}{
\strong{Deprecated Function} \tab \strong{Modern Function} \cr
\code{ AccountSummary } \tab \code{ \link{lc_AccountSummary}} \cr
\code{ AccountSummary_ } \tab \code{ \link{lc_AccountSummary}}  \cr
\code{ AddFunds } \tab \code{\link{lc_AddFunds}} \cr
\code{ AvailableCash } \tab \code{\link{lc_AvailableCash}} \cr
\code{ CreatePortfolio } \tab \code{\link{lc_CreatePortfolio}} \cr
\code{ Dashboard } \tab \code{\link{lc_Dashboard}} \cr
\code{ DetailedNotesOwned } \tab \code{\link{lc_DetailedNotesOwned}} \cr
\code{ MakeCredential } \tab \code{\link{lc_MakeCredential}} \cr
\code{ NotesOwned }\tab  \code{\link{lc_NotesOwned}} \cr
\code{ ListedLoans } \tab \code{\link{lc_ListedLoans}} \cr
\code{ SubmitOrder } \tab \code{\link{lc_SubmitOrder}} \cr
}
}
