% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClub.R
\docType{package}
\name{LendingClub}
\alias{LendingClub}
\alias{LendingClub-package}
\title{Lending Club}
\description{
Providing the tools to work with the Lending Club API.
Both the primary and secondary markets can be accessed
with this package.
To learn more about the LendingClub package, start with
the vignette:
\code{vignette("LendingClub")}
}
\author{
Ryan Kuhn, CPA
}

