% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClub.R
\docType{package}
\name{LendingClub}
\alias{LendingClub}
\alias{LendingClub-package}
\title{Lending Club}
\description{
Providing the tools to work with the Lending Club API.
Both the primary and secondary markets can be accessed
with this package.
}
\details{
Lending Club is a peer-to-peer lending service where loans
are broken up into $25 notes that investors buy with
expectation of earing a return on the interest. This
package is built to assist the investor manage
their account and maximize their return.

To learn more about the LendingClub package, start with
the vignette:
\code{vignette("LendingClub")}
}
\seealso{
\itemize{
\item{\url{http://ryankuhn.net/LendingClub}}
\item{\url{http://github.com/kuhnrl30/LendingClub/issues} for bug reports}
}
}
\author{
Ryan Kuhn, CPA
}
