% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCredential.R
\name{MakeCredential}
\alias{MakeCredential}
\title{Create the Lending Club credential}
\usage{
MakeCredential(investorID, APIkey)
}
\arguments{
\item{investorID}{User's ID from the Account Summary page}

\item{APIkey}{User's API key from the account settings page}

\item{...}{See the details section for more information.}
}
\description{
Creates the credential needed to access the other functions in this
 package.
}
\details{
This is the first step to accessing the API. Pass the function
your investor ID and API key.  The investor ID can be found on your
account summary page and the API Key can be found on your account
settings page. The function will create an object of class
LendingClub_credential. If the credential is stored to the variable
LC_CRED, then the other functions in the package will be able to
read the credential from the global environment.  Else, you may
pass the credential to the other functions as an argument.
}
\examples{
LC_CRED<- MakeCredential("investorID", "APIkey")

}
