% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubmitOrder.R
\name{SubmitOrder}
\alias{SubmitOrder}
\title{Order a note}
\usage{
SubmitOrder(loanId, amount, portfolioId = NULL)
}
\arguments{
\item{loanId}{unique ID for loans. These can be found using the
ListedLoans() function.}

\item{amount}{amount to be purchased}

\item{portfolioId}{unique ID for the portfolio. It can be found
using the PortfoliosOwned() function.}
}
\description{
Uses Lending Club's API to create an order to purchase a note
}

